Now using sync for Adding Liq to Well to lower gas
