+++ projects/sdk/src/lib/farm/actions/index.ts
@@ -16,7 +16,7 @@ import { ExchangeUnderlying } from "./ExchangeUnderlying";
 import { RemoveLiquidityOneToken } from "./RemoveLiquidityOneToken";
 import { WellSwap } from "./WellSwap";
 import { WellShift } from "./WellShift";
-import { WellAddLiquidity } from "./WellAddLiquidity"; 
+import { WellSync } from "./WellSync"; 
 import { DevDebug } from "./_DevDebug";
 
 export {
@@ -49,7 +49,7 @@ export {
   // DEX: Wells
   WellSwap,
   WellShift,
-  WellAddLiquidity,
+  WellSync,
 
   // Developers
   DevDebug
